ems grid problem solved

+++ js/script-ems.js
@@ -1,17 +1,12 @@
 // CREATE AN ARRAY OF EMPLOYEES
 let employees = []
+let count = 0
 
 // GET DOM ELEMENTS
-let empForm     = document.querySelector('#addForm')
-let empTable    = document.querySelector('#empTable')
-let empCount    = document.querySelector('#empCount')
+let empForm     = document.querySelector('#addForm');
+let empTable    = document.querySelector('#empTable');
+let empCount    = document.querySelector('#empCount');
 let empTableBody = document.querySelector('tbody')
-
-    let empID       = document.querySelector('#id')
-    let empName     = document.querySelector('#name')
-    let empExt      = document.querySelector('#extension')
-    let empEmail    = document.querySelector('#email')
-    let empDept     = document.querySelector('#department')
 
 // CHECK TO SEE IF STORAGE OBJECT EXISTS WHEN THE PAGE LOADS
 // IF DOES, RETURN STORAGE OBJECT INTO ARRAY INSTEAD OF POPULATED ARRAY
@@ -32,9 +27,11 @@ buildGrid();
 function addEmployee(employee) {
     // Add the new employee to the array
     employees.push(employee);
-
+    console.log(employees)
+    count ++
+    empCount.value = `(${count})`;
     // Build the grid
-    buildGrid();
+     buildGrid();
 
     // Store the array in storage
     localStorage.setItem('employees', JSON.stringify(employees));
@@ -55,6 +52,7 @@ empForm.addEventListener('submit', (e) => {
     const newEmployee = [empID, empName, empExt, empEmail, empDept]
     // PUSH THE NEW ARRAY TO THE *EXISTING* EMPLOYEES ARRAY
     addEmployee(newEmployee)
+    
     // BUILD THE GRID
     buildGrid()
     // RESET THE FORM
@@ -66,40 +64,45 @@ empForm.addEventListener('submit', (e) => {
 // DELETE EMPLOYEE
 empTable.addEventListener('click', (e) => {
     // CONFIRM THE DELETE
-    if (confirm('Are you sure you want to delete this employee?')) {
-        // GET THE SELECTED ROWINDEX FOR THE TR (PARENTNODE.PARENTNODE)
-
-        // REMOVE EMPLOYEE FROM ARRAY
-        empTable.deleteRow(e.target.parentElement.parentElement.rowIndex);
-        count--;
-
-        // BUILD THE GRID
-        //buildGrid();
+    if (e.target.classList.contains('delete-btn')) {
+        if (confirm('Are you sure you want to delete this employee?')) {
+            // CALL THE DELETEROW() METHOD TO DELETE SPECIFIC ROW IN TABLE
+            // PASS THE ROWINDEX FOR THE TR (PARENTNODE.PARENTNODE)
+            empTable.deleteRow(e.target.parentElement.parentElement.rowIndex);
+            // DECREMENT THE COUNTER
+            count--;
+            empCount.value = `(${count})`
+        }
     }
+        // BUILD THE GRID
+       // buildGrid();
+    
 
 });
 
 // BUILD THE EMPLOYEES GRID
 function buildGrid() {
     // REMOVE THE EXISTING SET OF ROWS BY REMOVING THE ENTIRE TBODY SECTION
-    empTableBody.innerHTML = '';
+    empTable.innerHTML = '';
 
     // REBUILD THE TBODY FROM SCRATCH
     const tbody = document.createElement('tbody');
 
     // LOOP THROUGH THE ARRAY OF EMPLOYEES
     // REBUILDING THE ROW STRUCTURE
-    console.log(employees)
+
     employees.forEach((employee) => {
         const row = document.createElement('tr');
-        row.innerHTML = `
-            <td>${employee.empID}</td>
-            <td>${employee.empName}</td>
-            <td>${employee.empExt}</td>
-            <td>${employee.empEmail}</td>
-            <td>${employee.empDept}</td>
-            <td><button class="btn btn-danger btn-sm delete-btn">X</button></td>
-        `;
+        console.log(employee[0])
+         row.innerHTML =  `
+                        <td>${employee[0]}</td>
+                        <td>${employee[1]}</td>
+                        <td>${employee[2]}</td>
+                        <td>${employee[3]}</td>
+                        <td>${employee[4]}</td>
+                        <td><button class="btn btn-danger btn-sm delete-btn">X</button></td>
+                    `;
+
         // Append the constructed row to the <tbody> tag
      tbody.appendChild(row);    
     });
@@ -110,7 +113,7 @@ function buildGrid() {
  empTable.appendChild(tbody);
 
  // UPDATE EMPLOYEE COUNT
- empCount.textContent = employees.length;
+ empCount.value = `(${count})`;
 
  // STORE THE ARRAY IN STORAGE
  localStorage.setItem('employees', JSON.stringify(employees));
